\name{mapGOtoData}
         \alias{mapGOtoData}
         \title{wrapper function to read cufflinks output Files}
         \description{This function parses the output files generated by the programs cufflinks, cuffcompare, cuffmerge and cuffdiff and generates a dataframe with all containing data}
         \usage{
         mapGOtoData(dataframe,genome,identifier,rowname,category)
         }
         \arguments{
           \item{dataframe}{dataframe where the GO Identifier will be added} 
           \item{genome}{which genome will be used}
           \item{identifier}{which type of identifiers will be used e.g. ensembl="ensgene",...} 
           \item{rowname}{name of the row that will be added}
           \item{category}{which ontologie will be used e.g. BP - biological process, CC - cellular component, MF - molecular function}
         }
         \examples{
         ## read cufflinks GTF File
         mapGOtoData(cuffdiff, "hg19","ensGene","BP","BP")
         }
         \seealso{\link{mapGOtoElement}, \link{mapAllGOtoData}, \link[goseq]{goseq}
         }