         \name{readCufflinksFiles}
         \alias{readCufflinksFiles}
         \title{reads cufflinks output Files}
         \description{This wrapper function parses the output files generated by the programs cufflinks, cuffcompare, cuffmerge and cuffdiff and generates a dataframe with all containing data}
         \usage{
         readCufflinksFiles(filename, program = "cufflinks")
         }
         \arguments{
           \item{filename}{path of the output file}
           \item{program}{output of which program have to be parsed                     (default="cufflinks")}
         }
         \examples{
         ## read cufflinks GTF File
         readCufflinksFiles(system.file("extdata","cufflinks.gtf",package="cuffGO"),"cufflinks")
         
         ## read cuffmerge GTF File
         readCufflinksFiles(system.file("extdata","cuffmerge.gtf",package="cuffGO"),"cuffmerge")
      
         ##read cuffcompare GTF File
         readCufflinksFiles(system.file("extdata","cuffcompare.gtf",package="cuffGO"),"cuffcompare")
         
         ## read cuffdiff File
         readCufflinksFiles(system.file("extdata","cuffdiff.diff",package="cuffGO"),"cuffdiff")
         }
        \details{
        the following files can be parsed with this function:
        \describe{
        \item{\code{transcripts.gtf}}{This GTF file contains Cufflinks' assembled isoforms}
        \item{\code{<outprefix>.combined.gtf}}{Cuffcompare reports a GTF file containing the "union" of all transfrags in each sample.}
        \item{\code{<outprefix>/merged.gtf}}{cuffmerge produces a GTF file that contains an assembly that merges together the input assemblies.}
        \item{\code{isoform_exp.diff}}{Transcript differential FPKM}
        \item{\code{gene_exp.diff}}{Gene differential FPKM. Tests differences in the summed FPKM of transcripts sharing each gene_id}
        \item{\code{tss_group_exp.diff}}{Primary transcript differential FPKM. Tests differences in the summed FPKM of transcripts sharing each tss_id}
        \item{\code{cds_exp.diff}}{Coding sequence differential FPKM. Tests differences in the summed FPKM of transcripts sharing each p_id independent of tss_id}
        }}
        \references{Cufflinks Manual:\url{http://cufflinks.cbcb.umd.edu/manual.html}}
        \value{
        \item{dataframe}{content of the Input file as dataframe}
        }
