\name{readTracking}
  \alias{readTracking}
  \title{reads output in tracking files generated by the programs like  cuffdiff and cufflinks}
  \description{This function reads output from the .fpkm_tracking, .count_tracking and .read_group_tracking files generated by cufflinks, cuffmerge and cuffdiff. The type of values in this files is changing with the kind of file (fpkm, count and read_group) and depends of the kind of analysis that is done. }
  \usage{readTracking(filename)}
  \arguments{
  \item{filename}{location of the cuffdiff output file in tracking format}
  }
  \value{
  \item{dataframe}{content of the parsed file as dataframe}
  }
  \format{
  Cufflinks, Cuffmerge and Cuffdiff calculates various tracking files that have different attributes. So the format of these files is not fixed, it changes with the type of tracking file that is read.
  }
  \references{Cufflinks Manual Cuffdiff Tracking Files: \url{http://cufflinks.cbcb.umd.edu/manual.html#cuffdiff_output}}
 